requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Jinja2==3.1.6
MarkupSafe==3.0.2
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
serverless-wsgi==3.0.5
openai==0.28
python-dotenv
twilio
requests
pydantic
supabase
PyPDF2
pinecone
pdfplumber
python-dateutil>=2.8.2
jinja2
weasyprint
boto3